Issue #21: Refactor project to modularize FHIR version-specific logic
