+++ bunsen-core/src/main/java/com/cerner/bunsen/codes/base/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Base classes for working with ConceptMaps and ValueSets. Consumers
+ * should generally not use this package, and instead consume the FHIR
+ * version-specific subclasses.
+ */
+package com.cerner.bunsen.codes.base;
