+++ bunsen-core/src/main/java/com/cerner/bunsen/codes/package-info.java
@@ -1,4 +1,6 @@
 /**
- * Classes for working with FHIR concept maps.
+ * Common for working with FHIR concept maps. Generally consumers will
+ * start with FHIR version-specific ConceptMaps and ValueSets classes
+ * that will produce the objects in this package.
  */
 package com.cerner.bunsen.codes;
