+++ bunsen-core/src/main/scala/com/cerner/bunsen/datatypes/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This package defines the mapping interface from Spark to FHIR data types for each version of FHIR
+ * supported.
+ */
+package com.cerner.bunsen.datatypes;
