+++ bunsen-stu3/pom.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
-  <artifactId>bunsen-core</artifactId>
+  <artifactId>bunsen-stu3</artifactId>
   <packaging>jar</packaging>
   <parent>
     <artifactId>bunsen-parent</artifactId>
@@ -9,8 +9,8 @@
     <version>0.2.3-SNAPSHOT</version>
     <relativePath>../pom.xml</relativePath>
   </parent>
-  <name>Bunsen Core</name>
-  <description>Encoder library for working with FHIR in Spark</description>
+  <name>Bunsen STU3</name>
+  <description>FHIR STU3 Support for Bunsen</description>
 
   <build>
     <plugins>
@@ -22,6 +22,14 @@
   </build>
 
   <dependencies>
+
+    <!-- Bunsen -->
+    <dependency>
+      <groupId>com.cerner.bunsen</groupId>
+      <artifactId>bunsen-core</artifactId>
+      <version>${project.version}</version>
+      <scope>compile</scope>
+    </dependency>
 
     <!-- Spark -->
     <dependency>
@@ -40,32 +48,19 @@
       <scope>provided</scope>
     </dependency>
 
-    <dependency>
-      <groupId>ca.uhn.hapi.fhir</groupId>
-      <artifactId>hapi-fhir-base</artifactId>
-      <scope>compile</scope>
-    </dependency>
-
-    <dependency>
-      <groupId>ca.uhn.hapi.fhir</groupId>
-      <artifactId>hapi-fhir-utilities</artifactId>
-      <scope>compile</scope>
-    </dependency>
-
-    <!-- Testing -->
-
-    <!-- HAPI FHIR DSTU3 resources included only for unit tests. -->
+    <!-- HAPI FHIR DSTU3 -->
     <dependency>
       <groupId>ca.uhn.hapi.fhir</groupId>
       <artifactId>hapi-fhir-validation-resources-dstu3</artifactId>
-      <scope>test</scope>
+      <scope>compile</scope>
     </dependency>
     <dependency>
       <groupId>ca.uhn.hapi.fhir</groupId>
       <artifactId>hapi-fhir-structures-dstu3</artifactId>
-      <scope>test</scope>
+      <scope>compile</scope>
     </dependency>
 
+    <!-- Testing -->
     <dependency>
       <groupId>junit</groupId>
       <artifactId>junit</artifactId>
