+++ bunsen-stu3/src/main/java/com/cerner/bunsen/stu3/python/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Support for Python-specific wrapper functions.
  */
-package com.cerner.bunsen.python;
+package com.cerner.bunsen.stu3.python;
