+++ bunsen-stu3/src/test/resources/xml/bundles/patient-1032702.fhir-bundle.xml
