+++ python/Makefile
@@ -2,6 +2,6 @@ init:
 	pip install -r requirements.txt
 
 test:
-	py.test tests
+	py.test -x tests
 
 .PHONY: init test
