+++ python/bunsen/stu3/__init__.py
@@ -1,0 +1,3 @@
+"""
+Bunsen STU3 Python API
+"""
